Refactored code for better separation of concerns
